Only allow running update on the main branch
Too many times I have run the command on a feature branch. Nothing
breaks permanently, but branches to be deleted should be compared to the
main branch, to see if they have been merged. Running the command in a
feature branch breaks this, but since the remote branch is purged
anyway, this leaves deleting local branches a manual operation.